requirements.txt
yfinance>=0.2.0
pyyaml>=6.0
requests>=2.28.0
